### 2026-02-16 #10063 @editorialbot
[PRE REVIEW]: BordAX: A High-Performance JAX Framework for Programmatic Reinforcement Learning
**Submitting author:** @kameronton (<a href="http://orcid.org/0009-0006-7495-9793">Roman Kniazev</a>)
**Repository:** https://github.com/SynthesisLab/bordax
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.1.0
