### 2026-04-23 #10063 @yewentao256
Hi  @HaoyuanHe0606, @tomzhengy would you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "BordAX: A High-Performance JAX Framework for Programmatic Reinforcement Learning".

Thank you so much!
Wentao
