### 2026-04-26 #10470 @HaoyuanHe0606
Hi @kameronton, thanks for the submission. BordAX is a clearly scoped piece of work and I enjoyed reading through it. The framework's modular split into collector, batch builder, and updater is genuinely well thought out, and the SB3 benchmark with five seeds and a checked in reproduction script makes the performance claim easy to verify. The git history goes back to January 2025, so this isn't a rushed out project, and the unit and integration tests give a reasonable safety net. That said, I think there are a few things worth tightening before acceptance.

The biggest thing that stood out is that the project still feels like solo work being prepared for publication rather than a software project that has invited the community in. The repository has no CONTRIBUTING file, no code of conduct, no issue or pull request templates, and no clearly stated support pathway, so a prospective user or contributor doesn't know how to file a bug, ask a question, or propose a change. Adding even a short CONTRIBUTING document and a sentence in the README pointing at GitHub Issues for support would address most of the JOSS community guidelines requirements without much effort. Relatedly, all four merged pull requests are from Dependabot, the issue tracker is empty, and there are no GitHub Releases or git tags. Cutting a tagged v0.1.0 release would give reviewers and downstream users something stable to pin to and would also be a natural place to attach a Zenodo DOI when the time comes.

I'd also encourage you to clarify the authorship picture. Nathanaël Fijalkow is listed as a coauthor on the paper, but the only commit attributed to him is the one that adds paper.md and paper.bib. JOSS is fine with non code contributions, but it helps to make those contributions visible somewhere, whether in an Acknowledgements paragraph, an AUTHORS file, or a sentence in the paper itself describing each author's role. While you're in the paper, I noticed that Gymnasium is currently cited via the OpenAI Gym entry (brockman2016openai), which conflates two separate projects. Gymnasium is the Farama Foundation fork and should be cited as such. The bib file also contains two entries for Lu et al. 2022 (lu2022discovered and gymnax_lu2022) and neither is actually referenced in the body, so they can be consolidated or dropped. The "Software design" heading is lowercase where JOSS asks for "Software Design", a small but easy fix.

On the testing and CI side, the README advertises 48 tests at 77 percent coverage as static badges, but I couldn't find a CI workflow that actually runs the test suite. The only GitHub Actions file in the repo is the JOSS PDF draft workflow, and it lives on the JOSS branch rather than main. Adding a small workflow that runs pytest on push and pull requests, and ideally publishing coverage to a service so the badge reflects current state, would substantially raise confidence that the project is being actively maintained. While you're at it, it might be worth merging the JOSS branch into main, or at least making it obvious to a reviewer landing on the default branch that the paper exists. Right now paper.md and paper.bib aren't visible without poking at the branch list.

The Research Impact Statement is the section I'd push on hardest from a JOSS perspective. The benchmark numbers are good, but realized impact evidence such as publications using BordAX, external integrations, or downstream users is absent, and the credible near term significance evidence is essentially the benchmark plus the open repository. If there are concrete near term plans, such as papers in preparation that depend on BordAX, courses or labs adopting it, or specific PRL benchmarks you intend to publish, weaving even a sentence or two of that into the statement would make it much more compelling. Documentation wise, the README is excellent for a high level orientation, but the in code docstrings are uneven across modules, and there's no rendered API reference. A lightweight Sphinx or MkDocs site published via GitHub Pages would go a long way and pairs naturally with the release tagging suggestion above. Finally, algorithms/base.py still contains commented out skeletons for A2C, DDQN, SAC, and TD3 that imply work in progress. If those aren't going to land soon, I'd remove them so readers aren't misled about the supported algorithm set.

Overall this is a solid submission with a real engineering core and a genuine niche in the programmatic RL space. None of the issues above are blocking, and most are quick fixes once you decide to do them. Thanks again for the work, and looking forward to the revision.
