### 2026-04-29 b229060
chore: add benchmark extra
