### 2026-04-29 eeed8f1
fix: wrap sb3 eval env with monitor
