### 2026-04-29 ac3d34d
docs: fix pytest setup command
