### 2026-04-29 9b6568d
chore: add benchmark extra
