### 2026-04-29 cbe3d5a
docs: fix dqn agent example
