### 2026-04-29 #10470 @tomzhengy
@kameronton Thank you for your submission! BordAX is a technically interesting project and nicely scoped. The core engineering looks great. I was able to run the main PPO/DQN functionality locally and also run the tests. I agree with @HaoyuanHe0606 's review - It would be great to add standard open-source/community files, clarification on your coauthors + contributions, and evidence of realized external impact such as downstream users, publications using BordAX, integrations, or community adoption.

On top of this I found several issues that should be addressed:

1. The documented test command did not work for me from a clean setup. After running `uv sync`, `uv run pytest tests/ -m "not slow" -v` failed because `pytest` was not installed by the default environment. After installing the dev extra, triggering pytest through the console entry point still failed to import `bordax` on my machine, while `uv run python -m pytest tests/ -m "not slow" -v` worked from the repository root. I opened https://github.com/SynthesisLab/bordax/pull/11 to address this.

2. The benchmark script required an undocumented extra dependency. Running `compare_sb3.py` failed in my clean environment until I manually installed `stable-baselines3`. I opened https://github.com/SynthesisLab/bordax/pull/12 to add a `benchmark` optional extra and update the README benchmark command.

3. I could not exactly reproduct the benchmark claim. I reproduced that BordAX is faster than SB3, but the speedups differed from the paper/README values. I recommend documenting the benchmark environment more precisely and adding the benchmark dependency to an optional extra or setup instructions, as I am unsure where the discrepancy lies.

3.1 In my local run of `compare_sb3.py` with 5 seeds and 51,200 timesteps, I measured:
- BordAX + Gymnax: 3.88s ± 0.06s, 13,180 steps/s, 2.8x faster than SB3
- BordAX + Gymnasium: 5.68s ± 0.04s, 9,010 steps/s, 1.9x faster than SB3
- Stable-Baselines3: 10.99s ± 0.07s, 4,658 steps/s
My run reproduced the direction of the claim but not the exact reported speedups.

4. When running the benchmark it emitted a Stable-Baselines3 warning that the evaluation environment was not wrapped with `Monitor`. I suggest either addressing that warning or explaining if it does/does not affect the comparison.

5. I noticed a small inconsistency in the example API documentation for DQN usage. The README includes an example using `make_agent("dqn", ...)`, but the implemented registry appears to use the more specific DQN agent key. I opened https://github.com/SynthesisLab/bordax/pull/13 to update the README policy-representation examples so they match the current public API.

Overall, I think the main algorithms and tests are in good shape, but I'd like to see the installation, testing, and benchmark reproduction path made exact before I check off the more of the reproducibility sections of JOSS items.
