### 2026-05-26 fa86bc7
update contribution guide and related templates
