+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug Report
+about: Report a reproducible bug
+labels: bug
+---
+
+## Description
+
+A clear description of what the bug is.
+
+## Steps to Reproduce
+
+```python
+# Minimal reproducible example
+```
+
+## Expected Behavior
+
+What you expected to happen.
+
+## Actual Behavior
+
+What actually happened. Include the full traceback if applicable.
+
+## Environment
+
+- OS:
+- Python version:
+- JAX version (`python -c "import jax; print(jax.__version__)"`):
+- BordAX version / commit:
