+++ paper.bib
@@ -1,0 +1,128 @@
+@software{jax2018github,
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url = {http://github.com/jax-ml/jax},
+  version = {0.3.13},
+  year = {2018}
+}
+
+@software{flax2020github,
+  author = {Jonathan Heek and Anselm Levskaya and Avital Oliver and Marvin Ritter and Bertrand Rondepierre and Andreas Steiner and Marc van {Z}ee},
+  title = {{F}lax: A neural network library and ecosystem for {JAX}},
+  url = {http://github.com/google/flax},
+  version = {0.5.0},
+  year = {2020}
+}
+
+@software{optax2020github,
+  author = {Matteo Hessel and David Budden and Fabio Viola and Mihaela Rosca and Eren Sezener and Tom Hennigan},
+  title = {Optax: composable gradient transformation and optimisation, in {JAX}},
+  url = {http://github.com/deepmind/optax},
+  version = {0.0.6},
+  year = {2020}
+}
+
+@software{gymnax2022github,
+  author = {Robert Tjarko Lange},
+  title = {{gymnax}: A {JAX}-based Reinforcement Learning Environment Library},
+  url = {http://github.com/RobertTLange/gymnax},
+  version = {0.0.4},
+  year = {2022}
+}
+
+@article{schulman2017proximal,
+  title = {Proximal Policy Optimization Algorithms},
+  author = {John Schulman and Filip Wolski and Prafulla Dhariwal and Alec Radford and Oleg Klimov},
+  journal = {arXiv preprint arXiv:1707.06347},
+  year = {2017},
+  url = {https://arxiv.org/abs/1707.06347}
+}
+
+@article{mnih2015human,
+  title = {Human-level control through deep reinforcement learning},
+  author = {Volodymyr Mnih and Koray Kavukcuoglu and David Silver and Andrei A. Rusu and Joel Veness and Marc G. Bellemare and Alex Graves and Martin Riedmiller and Andreas K. Fidjeland and Georg Ostrovski and Stig Petersen and Charles Beattie and Amir Sadik and Ioannis Antonoglou and Helen King and Dharshan Kumaran and Daan Wierstra and Shane Legg and Demis Hassabis},
+  journal = {Nature},
+  volume = {518},
+  number = {7540},
+  pages = {529--533},
+  year = {2015},
+  doi = {10.1038/nature14236}
+}
+
+@article{raffin2021stable,
+  title = {Stable-Baselines3: Reliable Reinforcement Learning Implementations},
+  author = {Antonin Raffin and Ashley Hill and Adam Gleave and Anssi Kanervisto and Maximilian Ernestus and Noah Dormann},
+  journal = {Journal of Machine Learning Research},
+  volume = {22},
+  number = {268},
+  pages = {1--8},
+  year = {2021},
+  url = {http://jmlr.org/papers/v22/20-1364.html}
+}
+
+@article{towers2026gymnasium,
+  title={Gymnasium: A standard interface for reinforcement learning environments},
+  author={Towers, Mark and Kwiatkowski, Ariel and Balis, John and De Cola, Gianluca and Deleu, Tristan and Goul{\~a}o, Manuel and Andreas, Kallinteris and Krimmel, Markus and Kg, Arjun and Perez-Vicente, Rodrigo and others},
+  journal={Advances in Neural Information Processing Systems},
+  volume={38},
+  year={2026}
+}
+
[…]
