+++ paper.md
@@ -1,0 +1,101 @@
+---
+title: 'BordAX: A High-Performance JAX Framework for Programmatic Reinforcement Learning'
+tags:
+  - Python
+  - JAX
+  - reinforcement learning
+  - programmatic policies
+  - decision trees
+  - interpretable machine learning
+authors:
+  - name: Roman Kniazev
+    orcid: 0009-0006-7495-9793
+    affiliation: 1
+  - name: Nathanaël Fijalkow
+    orcid: 0000-0002-6576-4680
+    affiliation: 1
+affiliations:
+  - name: CNRS, LaBRI, University of Bordeaux, France
+    index: 1
+date: 12 February 2026
+bibliography: paper.bib
+---
+
+# Summary
+
+BordAX is a Python framework for reinforcement learning (RL) built on JAX [@jax2018github].
+It provides a modular, fully JIT-compilable training pipeline that supports multiple policy representations, including standard multilayer perceptrons (MLPs), differentiable decision trees (DTSemNet), and boolean function networks (HyperBool).
+BordAX currently implements Proximal Policy Optimization (PPO) [@schulman2017proximal] and Deep Q-Networks (DQN) [@mnih2015human], and is designed so that additional algorithms can be composed from interchangeable collector, batch-builder, and updater components.
+The framework supports both Gymnax [@gymnax2022github] environments, which enable full JIT compilation of the training loop via `jax.lax.scan`, and standard Gymnasium [@towers2026gymnasium] environments.
+On a CartPole-v1 benchmark with identical hyperparameters, BordAX with Gymnax achieves approximately 2.8 times higher throughput than Stable-Baselines3 [@raffin2021stable].
+
+# Statement of need
+
+Programmatic Reinforcement Learning (PRL) is a subfield of RL concerned with learning structured, interpretable policies such as decision trees, boolean expressions, or symbolic programs, rather than opaque neural networks [@verma2018programmatically; @landajuela2021discovering].
+Researchers in PRL often need to train and evaluate several types of policy representations under the same algorithmic conditions.
+Existing RL frameworks are typically built around a fixed neural network policy and do not provide straightforward mechanisms for swapping in non-standard architectures like differentiable decision trees [@silva2020optimization] or boolean networks.
+
+Stable-Baselines3 [@raffin2021stable] is a widely used PyTorch-based RL library, but its reliance on Python-level iteration and PyTorch's eager execution model limits throughput.
+PureJaxRL [@purejaxrl2023] demonstrated that implementing the entire RL training loop in JAX and compiling it end-to-end via XLA yields large speedups, but it is structured as a collection of standalone scripts rather than a reusable library with modular components.
+Brax [@brax2021github] provides JIT-compiled environments and training but is focused on continuous control in simulated physics rather than serving as a general-purpose RL framework with pluggable policy types.
+
+BordAX addresses these gaps by providing:
+
+- A modular architecture in which the algorithm is defined as the composition of a data collector, a batch builder, and a parameter updater, making it straightforward to implement new algorithms or modify existing ones.
+- First-class support for non-neural policy representations (differentiable decision trees and boolean function networks) alongside standard MLPs, enabling controlled comparisons across policy types.
+- Full JIT compilation of the training loop when using Gymnax environments, and JIT compilation of the update step when using Gymnasium environments.
+- A clean functional design compatible with JAX transformations (`jax.jit`, `jax.vmap`, `jax.grad`).
+
+The target audience is researchers working on programmatic or interpretable reinforcement learning who need a fast, modular framework for experimenting with different policy representations and algorithms.
+
+# State of the field
+
+Several frameworks exist for reinforcement learning in Python.
+
[…]
