### 2026-05-26 df966d3
add docs stub
