+++ docs/api/environments.md
@@ -1,0 +1,3 @@
+# bordax.environments
+
+::: bordax.environments.utils
