+++ docs/api/training.md
@@ -1,0 +1,7 @@
+# bordax.training
+
+::: bordax.training.trainer
+
+::: bordax.training.updaters
+
+::: bordax.training.evaluation
