+++ docs/api/algorithms.md
@@ -1,0 +1,7 @@
+# bordax.algorithms
+
+::: bordax.algorithms.base
+
+::: bordax.algorithms.losses
+
+::: bordax.algorithms.utils
