### 2026-05-26 baf93b6
add docstrings throughout the repo
