+++ bordax/environments/utils.py
@@ -30,27 +30,74 @@ class EnvParams:
 
 
 class EnvAdapter(ABC):
+    """Unified interface over Gymnax and Gymnasium environments.
+
+    All environments are treated as vectorised: ``reset`` and ``step``
+    operate on a batch of ``num_envs`` parallel episodes.
+
+    Attributes:
+        is_jittable: Whether the environment can be used inside
+            ``jax.jit`` / ``jax.lax.scan``. ``True`` for Gymnax,
+            ``False`` for Gymnasium.
+        num_envs: Number of parallel environment instances.
+        env: The underlying environment object.
+        env_params: Static environment parameters (e.g. episode length).
+    """
+
     is_jittable: bool
     num_envs: int
     env: Any
     env_params: Any
 
     @abstractmethod
-    def reset(self, key: PRNGKey) -> Tuple[EnvObs, EnvState]: ...
+    def reset(self, key: PRNGKey) -> Tuple[EnvObs, EnvState]:
+        """Reset all environments and return initial observations.
+
+        Args:
+            key: JAX random key used to seed the reset.
+
+        Returns:
+            Tuple of ``(obs, state)`` with leading batch dimension
+            ``num_envs``.
+        """
+        ...
 
     @abstractmethod
     def step(
         self, key: PRNGKey, state: EnvState, action: Any
-    ) -> Tuple[Any, EnvState, float, bool, Mapping[str, Any]]: ...
+    ) -> Tuple[Any, EnvState, float, bool, Mapping[str, Any]]:
+        """Step all environments forward by one timestep.
+
+        Args:
+            key: JAX random key for stochastic transitions.
+            state: Current environment state (batch of ``num_envs``).
+            action: Actions to apply, shape ``(num_envs, ...)``.
+
+        Returns:
+            Tuple of ``(obs, state, reward, done, info)``.
+        """
+        ...
 
     @abstractmethod
-    def action_space(self) -> Space: ...
+    def action_space(self) -> Space:
+        """Return the single-environment action space."""
+        ...
 
     @abstractmethod
-    def obs_space(self) -> Space: ...
+    def obs_space(self) -> Space:
+        """Return the single-environment observation space."""
+        ...
 
 
 class EnvGymnaxAdapter(EnvAdapter):
+    """Adapter for Gymnax environments (fully JIT-compilable).
+
+    Wraps a Gymnax environment with ``jax.vmap`` across ``num_envs``
+    parallel instances. Both ``reset`` and ``step`` are JIT-compiled.
+    The entire training loop can be compiled end-to-end when using this
+    adapter with an on-policy algorithm.
+    """
+
     def __init__(self, env_name: str, env_config, num_envs: int = 1):
         self.is_jittable = True
         self.num_envs = num_envs
@@ -87,6 +134,13 @@ class EnvGymnaxAdapter(EnvAdapter):
 
 
 class EnvGymnasiumAdapter(EnvAdapter):
+    """Adapter for standard Gymnasium environments (non-JIT).
+
+    Wraps a vectorised Gymnasium environment (``gymnasium.make_vec``).
+    Only the gradient update step can be JIT-compiled; environment
+    stepping runs in Python.
+    """
+
     def __init__(self, env_name: str, env_config, num_envs: int = 1):
         self.is_jittable = False
         self.num_envs = num_envs
@@ -118,6 +172,24 @@ class EnvGymnasiumAdapter(EnvAdapter):
 
 
 def make_env(env_name: str, env_config, num_envs: int = 1) -> EnvAdapter:
+    """Create a vectorised environment adapter by name.
+
+    Args:
+        env_name: Environment identifier with a backend prefix, e.g.
[…]
