### 2026-05-26 9e3128b
remove non-implemented algos
