+++ paper.bib
@@ -1,9 +1,15 @@
 @software{jax2018github,
-  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Yash Katariya and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+
   url = {http://github.com/jax-ml/jax},
+
   version = {0.3.13},
-  year = {2018}
+
+  year = {2018},
+
 }
 
 @software{flax2020github,
@@ -86,6 +92,21 @@
   organization = {PMLR}
 }
 
+@inproceedings{panda2024vanilla,
+  author       = {Subrat Prasad Panda and
+                  Blaise Genest and
+                  Arvind Easwaran and
+                  Ponnuthurai Nagaratnam Suganthan},
+  title        = {Vanilla Gradient Descent for Oblique Decision Trees},
+  booktitle    = {{ECAI} 2024 - 27th European Conference on Artificial Intelligence},
+  series       = {Frontiers in Artificial Intelligence and Applications},
+  volume       = {392},
+  pages        = {1140--1147},
+  year         = {2024},
+  url          = {https://doi.org/10.3233/FAIA240607},
+  doi          = {10.3233/FAIA240607},
+}
+
 @article{topin2021iterative,
   title = {Iterative Bounding {MDP}s: Learning Interpretable Policies via Non-Interpretable Methods},
   author = {Nicholay Topin and Stephanie Milani and Fei Fang and Manuela Veloso},
@@ -93,7 +114,8 @@
   volume = {35},
   number = {11},
   pages = {9923--9931},
-  year = {2021}
+  year = {2021},
+  doi = {10.1609/aaai.v35i11.17192}
 }
 
 @article{landajuela2021discovering,
