### 2026-06-25 826dafd
fix citations
